pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "ai-career-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "python-docx>=1.2.0",
    "gradio>=6.11.0",
    "ipykernel>=7.2.0",
    "openai>=2.30.0",
    "openai-agents>=0.13.6",
    "pypdf>=5.0.0",
    "python-dotenv>=1.2.2",
    "reportlab>=4.0.0",
]

[project.scripts]
career-assistant = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["ui", "utils", "career_agents", "agent_tools"]

[tool.hatch.build.targets.wheel.force-include]
"ui/html" = "ui/html"

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"

[dependency-groups]
dev = [
    "markdown>=3.10.2",
    "playwright>=1.58.0",
]
